The fetch_mail feature shifted from ndf_tags to views file
