+++ gnowsys-ndf/gnowsys_ndf/ndf/templates/ndf/mailclient.html
@@ -2,7 +2,6 @@
 {% load i18n %}
 {% load pagination_tags %}
 {% load ndf_tags %}
-{% load get_mails_in_box from ndf_tags %}
 {% block title %} {{title}} {% endblock %}
